requirements.txt
asgiref==3.2.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
static3==0.7.0
gunicorn==20.0.0
psycopg2==2.8.4
djangorestframework==3.11.0